Check what wld.renderer is not null
It would segfault for me, turns out wld.renderer is null, not entirely
sure why yet but add a check for that so it'll exit instead of
segfaulting.

Note that using perror() instead of just exiting gives at least *some*
info as to what's going wrong:

	Can't create renderer: Permission denied

So maybe add that to die()? But that's a larger change so didn't add
that here.